+++ RVCalendarWeekView/Lib/MSCollectionViewCalendarLayout/MSCollectionViewCalendarLayout.m
@@ -1082,7 +1082,7 @@ NSUInteger const MSCollectionMinBackgroundZ = 0.0;
 
 -(void)scrollCollectionViewToCurrentTime:(BOOL)animated{
     //int hour  = [NSDate now:@"device"].hour;    //Why does this crash..
-    int hour    = [NSDate parse:NSDate.date.toDateTimeString timezone:@"device"].hour;
+    NSUInteger hour    = [NSDate parse:NSDate.date.toDateTimeString timezone:@"device"].hour;
     CGFloat y   = MIN(hour*self.hourHeight, self.collectionView.contentSize.height - self.collectionView.frame.size.height);
     [self.collectionView setContentOffset:CGPointMake(self.collectionView.contentOffset.x, y) animated:animated];
 }
